Fix upload handler

+++ product-service/internal/model/product.go
@@ -1,13 +1,13 @@
 package model
 
 type Product struct {
-	ID          int
-	SellerID    int    `json:"user_id"`
-	Title       string `json:"title"`
-	Description string `json:"description"`
-	Category    string `json:"category"`
-	Price       int    `json:"price"`
-	Quantity    int    `json:"quantity"`
+	ID          int64
+	SellerID    int64   `json:"user_id"`
+	Title       string  `json:"title"`
+	Description string  `json:"description"`
+	Category    string  `json:"category"`
+	Price       float64 `json:"price"`
+	Quantity    int64   `json:"quantity"`
 	Images      []ProductImage
 }
 type ProductImage struct {
